requirements.txt
# Core
Django==5.0.1
django-environ==0.11.2
gunicorn==21.2.0
psycopg2==2.9.10

# For the persistence stores
psycopg2-binary==2.9.10

# Rest apis
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
whitenoise==6.6.0

# Certificates
cryptography==41.0.7

#setup
setuptools==75.2.0

#gcp
google-cloud-storage==2.18.2

# all-auth
django-allauth==65.0.2

#authlib
Authlib==1.3.2

# django-silk==5.3.1

channels==4.2.0
daphne==4.1.2
